% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignNicknames.R
\name{assignNicknames}
\alias{assignNicknames}
\title{Assign nicknames}
\usage{
assignNicknames(data, string = "Animal")
}
\arguments{
\item{data}{typically - a "track" object., but really, any data frame with an "id" column.}

\item{string}{a basic string for a nickname, like "Moose"}
}
\value{
the data, with an additional column called "nickname" with the numbered short name.
}
\description{
Most of the ID's of individuals are extremely unwieldly multi-digit codes.  This function gives nicer (more memorable) numbered short codes ... like "Moose1", "Moose2", etc.
}
