% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEnvDataGrid.R
\name{createEnvDataGrid}
\alias{createEnvDataGrid}
\title{Create gridded data for querying Movebank Env-Data}
\usage{
createEnvDataGrid(
  lats,
  lons,
  start,
  finish = start,
  dt = 0,
  savefile = TRUE,
  fileout = NULL
)
}
\arguments{
\item{dt}{time interval between desired time points.}

\item{savefile}{whether to save a csv file}

\item{fileout}{name of csv file to save}

\item{{lats, lons}}{vectors of latitude and longitude for the desired grid.  Can also be a single location.}

\item{{start, finish}}{vectors of POSIX times of start and finish for the query.}
}
\value{
Either nothing (if csv file saved) or the character string data frame with correct formatting.
}
\description{
Takes a vector of latitudes and longitudes (typically evenly spaced) and a vector of times and generates the strictly formatted data frame needed to upload to Env-Data to obtain covariates for movebank.  Essentially, an R-based substitute for the "gridded data" tool.
}
\examples{
 lats <- seq(38.8, 39.0, length = 40)
 lons <- seq(-77.12, -76.91, length = 40) 
 start <- ymd("2014-01-01")
 finish <- ymd("2014-12-31")
 dt <- ddays(20)
 example <- createEnvDataGrid(lats, lons, start, finish, dt, savefile = FALSE)
 head(example)

}
\seealso{
createEnvDataRequest, uploadEnvDataRequest, createEnvDataDataFrame
}
