% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadEnvData.R
\name{downloadEnvData}
\alias{downloadEnvData}
\title{Download and return annotated data from EnvData}
\usage{
downloadEnvData(header, returndata = TRUE, filename = NULL)
}
\arguments{
\item{header}{the header file outputted by \code{\link{uploadEnvDataRequest}}}

\item{returndata}{whether to return the data directly}

\item{filename}{filename (and location) to save annotated data to.  If left empty, user will be prompted.}
}
\value{
the annotated data frame
}
\description{
Download and return annotated data from EnvData
}
