% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMultiMigration.R
\name{fitMultiMigration}
\alias{fitMultiMigration}
\title{Fit multiple migrations}
\usage{
fitMultiMigration(data, span1, span2, plot = TRUE)
}
\arguments{
\item{data}{A processed daily averaged data frame, i.e. output of \code{\link{processMovedata}}}

\item{span1}{A vector of *initial* times for analysis windows.  Should be of length *k*, where *k* is the number of migrations to fit.}

\item{span2}{A vector of *final* times for analysis windows}

\item{plot}{Whether or not to visualize the output with \code{\link{plotMultimigrationFit}}}
}
\value{
a data frame with the date, duration, estimated ranging areas of each of the migrations.
}
\description{
This function allows for the analysis of multiple migration in a single movement time series.
}

