% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVT.R
\name{getVT}
\alias{getVT}
\title{Generate vector statistics}
\usage{
getVT(track, ...)
}
\arguments{
\item{track}{a track object generated using \code{processMovedata}}

\item{units}{one of ('secs', 'mins', 'hours', 'days') used to identify the unit of time for output (default = 'hours').}

\item{geoDist}{to estimate step lengths using the x/y coordinate units (FALSE) or 
the WGS ellipsoidal distances using LonLat (uses the function pointDistance(..., lonlat=TRUE) from raster package).}
}
\value{
Returns the original data frame with "z", "z.start", "z.end" (complex numbers indicating location), 
'stepLength' (step length), 'phi' and 'theta' (absolute and relative turning angles, respectively), 't.start', 
't.end', 't.mid' (numeric time), dT (difference in time over step), 'v' (velocity, default meters / hour), and
't.mid.POSIX' (time mid point in POSIX).
}
\description{
Get VT
}
\details{
Generates vector (i.e., movement) statistics for a track object.
}
\examples{
## Not run: 
### Enter Login and Password information for Movebank
# login <- movebankLogin(username="xxxx", password="xxxx")

# Download data
ds <- getMovebankData(study="ABoVE: NPS Dall Sheep Lake Clark", login=login) 

# Process data for use in above package
ds.mt <- processMovedata(ds,  idcolumn = "individual_id", projTo = '+init=epsg:3338', 
                        keepCols = c('sex'), dailymean = TRUE)

# Summarize and find individual
ds.vt <- getVT(ds.mt, units = 'hours', geoDist = FALSE)
}
\seealso{
\link{processMovedata}, \link{pointDistance}
}
