% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof}
\alias{gof}
\title{Goodness of Fit}
\usage{
gof(do, fits, obs, flag = NULL)
}
\arguments{
\item{do}{the 'test' data set used in cross-validation procedures.}

\item{fits}{the predicted values derived from a training model with test data as inputs.}

\item{obs}{the predicted values derived from a test model with test data as inputs.}

\item{flag}{a flag passed for whether a model converged (0 = TRUE, 1 = FALSE), usually as a result of insufficient test data.}
}
\value{
Returns a list containing the fitted and observed values (obj[[i]]$values) and cross-validation metrics (obj[[i]]$cv).
}
\description{
An internal function for estimating correlation metrics for cross-validation procedures
}
\author{
Peter Mahoney
}
