% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lio.clogit.R
\name{lio.clogit}
\alias{lio.clogit}
\title{Leave-individuals-out for SSF}
\usage{
lio.clogit(
  dat,
  form,
  thin = NULL,
  cluster = "ID_year",
  ID = "AnimalID",
  inPar = TRUE,
  ncores = NULL
)
}
\arguments{
\item{dat}{a movetrack object with pseudo-absences points (.e., available; Used = 0) and used points (Used = 1) stratified by step and annotated with representative covariates.}

\item{form}{a formula object with the clogit formula to be evaluated using LIO cross-validation.}

\item{thin}{an integer representing the amount of thinning required before performing the cross-validation (not recommended and potentially biases results).}

\item{cluster}{a character string for the name of the clustering variable (e.g., 'ID_year').}

\item{ID}{a character string for the name of the animal IDs. Used to partition individuals for the LIO procedure.}

\item{inPar}{Boolean for whether or not procedure should be performed in parallel (TRUE by default).}

\item{ncores}{if inPar = TRUE, ncores will define the number of cpu cores to use.}
}
\value{
Returns a list of lists indexed by individual.  Nested within each individual's element is a second list containing the fitted and observed values (obj[[i]]$values) and cross-validation metrics (obj[[i]]$cv).
}
\description{
A function for generating leave-individuals-out cross-validation output for a step-selection model.
}
\seealso{
\link{processMovedata}
}
\author{
Peter Mahoney
}
