% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSnowData24km.R
\name{loadSnowData24km}
\alias{loadSnowData24km}
\title{NSIDC Snow Model Functions}
\usage{
loadSnowData24km(year, day, filename = NULL, directory = ".",
  plot = FALSE, northamerica = TRUE, water = FALSE)
}
\arguments{
\item{plot}{whether to plot the map.}

\item{northamerica}{whether to limit to North America}

\item{water}{whether or not to include the open water category}

\item{{year, day}}{of data to download}

\item{{filename, directory}}{of the ascii raw snow file if raw data are already downloaded}

\item{...}{additional arguments to pass to polygon function}
}
\value{
a data frame that contains:  
\describe{
  \item{snow}{snow coverage: 1 = open water, 2 = no snow, 3 = sea ice, 4 = snow}
  \item{area}{of grid cell, in km^2}
  \item{lon,lat}{longitude and latitude - midpoint of grid cell}
  \item{lon.ll,lon.ul,lon.ur,lon.lr}{longitudes of corners of grid cell }
  \item{lat.ll,lat.ul,lat.ur,lat.lr}{latitudes of corners of grid cell}
}
}
\description{
This function downloads 24km x 24km snow ice data from the NSIDS repository (\link{ftp://sidads.colorado.edu/pub/DATASETS/NOAA/G02156/}) and processes the data.  An additonal function \code{plotSnowData} conveniently maps the output.
}
\examples{
#  The output is equivalent to the saved examples in \\code{\\link{snow24km}}
snow.data <- loadSnowData24km(2011, 16)
plotSnowData(snow.data); title("Jan 16, 2011")

}
\references{
National Ice Center. 2008, updated daily. IMS Daily Northern Hemisphere Snow and Ice Analysis at 1 km, 4 km, and 24 km Resolutions, Version 1. [Indicate subset used]. Boulder, Colorado USA. NSIDC: National Snow and Ice Data Center. doi: http://dx.doi.org/10.7265/N52R3PMC. [Date Accessed].
}
