% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.r
\docType{data}
\name{lonlat24km}
\alias{lonlat24km}
\title{24km Lat-Lon data}
\format{Two 1024 x 1024 matrices: \code{lon.matrix} and \code{lat.matrix} - of the longitudes and latitudes, respectively, of the snow grid.   A vector of \code{xs} and \code{ys} - coordinates that map the longituded and latitude to a unit circle.  Finally, a data frame \code{lonlat.df} which contains the midpoint, ll, ul, ur, and lr corners of each of the 1024 x 1024 = 1048576 lat-lon polygons}
\source{
National Ice Center. 2008, updated daily. IMS Daily Northern Hemisphere Snow and Ice Analysis at 1 km, 4 km, and 24 km Resolutions, Version 1. Boulder, Colorado USA. NSIDC: National Snow and Ice Data Center. doi: \link{http://dx.doi.org/10.7265/N52R3PMC}.
}
\usage{
data(lonlat24km)
}
\description{
The NCSID 24 km snow coverage products is provided as an ascii file with 1024 rows x 1024 columns using a polar projection centered around the north pole and covering the northern hemisphere.  The actual latitude and longitude coordinates of the data are provided separately as bin files - which are converted to a matrix here.
}
\details{
The raw lat-lon bin files are posted by NSIDC here: \link{ftp://sidads.colorado.edu/pub/DATASETS/NOAA/G02156/metadata/}
}
\examples{
data(lonlat24km)
dim(lat.matrix)
dim(lon.matrix)
require(fields)
image.plot(xs, ys, lat.matrix, main = "latitude")
image.plot(xs, ys, lon.matrix, main = "longitude")
}
