% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.movetrack.R
\name{map.movetrack}
\alias{map.movetrack}
\title{ggmap of a track}
\usage{
map.movetrack(movetrack, zoom = 6, ...)
}
\arguments{
\item{movetrack}{a \code{movetrack} class object, or any data-frame that contains (at least) three columns labeled "time", "x" and "y"}

\item{zoom}{the zoom level, an integer from 3 (continent) to 21 (building)}

\item{...}{additional parameters to pass to \code{\link{ggmap}}}
}
\description{
This is a quick wrapper around the \code{\link{ggmap}} function that plots a movetrack object (including with multiple individuals) and
}

