% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.track.R
\name{map}
\alias{map}
\title{ggmap of a track}
\usage{
map(movetrack, zoom = 6, individually = FALSE, id = NULL,
  exportPDF = FALSE, ...)
}
\arguments{
\item{movetrack}{a \code{movetrack} class object, or any data-frame that contains (at least) three columns labeled "time", "x" and "y"}

\item{zoom}{the zoom level, an integer from 3 (continent) to 21 (building)}

\item{individually}{whether to produce plots individually (TRUE) or as a population (FALSE; default).}

\item{exportPDF}{a logical indicating whether or not to export plots as a PDF (default = FALSE).}

\item{...}{additional parameters to pass to \code{\link{ggmap}}}
}
\value{
if exportPDF = TRUE, will return a single PDF of all individuals or a population to your working directory.
}
\description{
This is a quick wrapper around the \code{\link{ggmap}} function that plots a \code{track} object (including with multiple individuals) and
}
\examples{
### Enter Login and Password information for Movebank

## Not run: 
# login <- movebankLogin(username="xxxx", password="xxxx")

# Pull data from movebank
ge1 <- getMovebankData(study="ABoVE: HawkWatch International Golden Eagles", login=login)     #multiple birds

# Estimate daily mean locations
ge1.dailymean <- processMovedata(ge1,  id = "individual_id", dailymean = TRUE)


# one bird, three deployments
ge2 <- getMovebankData(study="Aquila chrysaetos interior west N. America, Craigs, Fuller", animalName="629-26704", login=login) 

# Estimate daily mean locations
ge2.dailymean <- processMovedata(ge2,  id = "individual_id", dailymean = FALSE)

# Plot daily movements (currently only displays 1 individual)
plot(ge2.dailymean)
}
