% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.r
\docType{data}
\name{Oscillations}
\alias{Oscillations}
\title{Oscillations}
\format{data frame with 7 columns:
\describe{
  \item{year}{beginning 1950}
  \item{month}{1-12}
  \item{PDO}{indices are roughly standard normal ...}
  \item{AO}{}
  \item{NAO}{}
  \item{season}{October:March = winter;  April:September = summer}
  \item{Y2}{assigns January:March to the previous year}
}}
\usage{
data(Oscillations)
}
\description{
Monthly averages of large-scaled climate oscillations in the northern hemisphere: Pacific Decadal (PDO), Arctic (AO) and North Atlantic (NAO) from January 1950 to December 2017.
}
\examples{
data(Oscillations)
require(gridExtra)
ggplot.osc <- function(data = Oscillations, var, title = ""){
  ggplot(data \%>\% mutate(year = factor(year)), 
         aes(year, get(var))) + 
    geom_boxplot() + 
    ggtitle(title) + ylim(-2.5,2.5)
}

o <- subset(Oscillations, year >= 1990 & season == "winter") \%>\% mutate(year = Y2)
pdo <- ggplot.osc(o, "PDO", "Pacific Decadal Oscillation")
ao <- ggplot.osc(o, "AO", "Arctic Oscillation")
nao <- ggplot.osc(o, "NAO", "North Atlantic Oscillation")
grid.arrange(pdo, ao, nao,  nrow = 3)
}
\references{
PDO: \link{http://research.jisao.washington.edu/pdo/}; AO: \link{http://www.cpc.noaa.gov/products/precip/CWlink/daily_ao_index/ao_index.html}; NAO: \link{https://www.ncdc.noaa.gov/teleconnections/nao/}
}
