% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.track.R
\docType{methods}
\name{plot,trackSPDF,missing-method}
\alias{plot,trackSPDF,missing-method}
\title{Quick plot of a track (against x-y-time)}
\usage{
\S4method{plot}{trackSPDF,missing}(x, y = NULL, id = NULL, layout = NULL,
  auto.par = NULL, exportPDF = FALSE, ...)
}
\arguments{
\item{x}{is a \code{AotM} or \code{AotM_df} class object, or any data-frame that contains (at least) three columns labeled "time", "x" and "y"}

\item{id}{id of animal(s) to plot.  If NULL - use subset or first animal.}

\item{exportPDF}{a logical indicating whether or not to export a PDF of all animal tracks to the local working directory (default=FALSE).}

\item{...}{options to be passed to plot functions}
}
\description{
Plotting x-y, time-x, time-y scan of a track.  This function will take x, y, and time coordinates or a \code{track} class object
}
