% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lio.R
\name{plot.lio}
\alias{plot.lio}
\title{Plot lio}
\usage{
plot.lio(lio, numpoints)
}
\arguments{
\item{lio}{An 'lio' class object.}

\item{numpoints}{Number of observed points to superimpose on density plots.}
}
\value{
Plots the fit of the with-held individual (test data, observed fit) again the remaining sampled individuals (training data, predicted fit). a data frame with median cross-validation metrics for all individuals in a leave-individual-out object.
}
\description{
A function for plotting observed vs. predicted values from a leave-individuals-out cross-validation procedure.
}
\author{
Peter Mahoney
}
