% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.movetrack.R
\name{plot.movetrack}
\alias{plot.movetrack}
\title{Quick plot of a track (against x-y-time)}
\usage{
plot.movetrack(movetrack = NULL, id = NULL, time, x, y = NULL,
  layout = NULL, auto.par = NULL, ...)
}
\arguments{
\item{movetrack}{a \code{movetrack} class object, or any data-frame that contains (at least) three columns labeled "time", "x" and "y"}

\item{...}{options to be passed to plot functions}
}
\description{
Plotting x-y, time-x, time-y scan of a track.  This function will take x, y, and time coordinates or a \code{track} class object
}

