% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMultimigrationFit.R
\name{plotMultiMigration}
\alias{plotMultiMigration}
\title{Plot Multiple Migration Fit}
\usage{
plotMultiMigration(data, M.summary, ...)
}
\arguments{
\item{data}{A processed daily averaged data frame, i.e. output of \code{\link{processMovedata}}}

\item{M.summary}{A data frame with the multi-migration summary (output of \code{\link{fitMultiMigration}})}
}
\description{
This function plots a multiple migration fit using the summary output of \code{\link{fitMultiMigration}}.
}

