% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickUTM.R
\name{quickUTM}
\alias{quickUTM}
\title{Quick UTM}
\usage{
quickUTM(lon, lat)
}
\arguments{
\item{{lon, lat}}{self-explanatory}
}
\value{
data frame with columns X and Y in automatically selected zone.
}
\description{
Uses the \code{\link{convUL}} function in \code{\link{PBSmapping}} to convert longitude and latitude to UTM.  This funciton is buried in \code{\link{processMovedata}}
}
