% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSpans.R
\name{selectSpans}
\alias{selectSpans}
\title{Select spans for multimigration analysis}
\usage{
selectSpans(time, x, y)
}
\arguments{
\item{{time, x, y}}{migration data}
}
\description{
Select spans for multimigration analysis
}
\details{
This function provides an interactive interface for picking out analysis windows for a multi migration analysis. Basically - follow the instructions at the top of the figure
}
