% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.r
\docType{data}
\name{snow24km}
\alias{snow24km}
\title{Sample 24km snow data frames}
\format{data frame with 11 columns:
\describe{
  \item{lon,lat}{longitude and latitude - midpoint of grid cell}
  \item{lon.ll,lon.ul,lon.ur,lon.lr}{longitudes of corners of grid}
  \item{lat.ll,lat.ul,lat.ur,lat.lr}{latitudes of corners of grid}
  \item{snow}{snow coverage: 2 = no snow, 3 = sea ice, 4 = snow}
  \item{area}{of grid cell, in km^2}
}}
\source{
National Ice Center. 2008, updated daily. IMS Daily Northern Hemisphere Snow and Ice Analysis at 1 km, 4 km, and 24 km Resolutions, Version 1. Boulder, Colorado USA. NSIDC: National Snow and Ice Data Center. doi: \link{http://dx.doi.org/10.7265/N52R3PMC}.
}
\usage{
data(snow24km)
}
\description{
Two data frames: \code{snow.2011.016} and \code{snow.2011.150} containing grid coordinates and snow coverage in North America from two days in 2011: Jan 16 (day 016) and May 30 (day 150).   The raw data files are located here: \link{ftp://sidads.colorado.edu/pub/DATASETS/NOAA/G02156/24km/2011/}
}
\examples{
data(snow24km)
par(mfrow = c(1,2))
plotSnowData(snow.2011.016); title("Jan 16, 2011")
plotSnowData(snow.2011.150); title("May 30, 2011")

}
