% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.lio.R
\name{summary.lio}
\alias{summary.lio}
\title{Summary output from a leave-individual-out object}
\usage{
\method{summary}{lio}(lio)
}
\arguments{
\item{lio}{An 'lio' class object.}
}
\value{
Returns a data frame with median cross-validation metrics for all individuals in a leave-individual-out object.
}
\description{
A summary function for leave-individual-out cross validation output
}
\author{
Peter Mahoney
}
