% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.movetrack.R
\name{summary.movetrack}
\alias{summary.movetrack}
\title{Summary method for a movetrack}
\usage{
\method{summary}{movetrack}(movetrack)
}
\arguments{
\item{movetrack}{output of \code{\link{processMovedata}}}
}
\value{
a data frame with a column of id's, number of observations, start time, duration and median observation interval of the track.
}
\description{
Summary method for a movetrack
}

