% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadEnvDataRequest.R
\name{uploadEnvDataRequest}
\alias{uploadEnvDataRequest}
\title{Upload an Env-Data request for Annotation}
\usage{
uploadEnvDataRequest(csv.file, xml.file, login = NULL, pw = NULL,
  link.file = "annotation_status_link.url")
}
\arguments{
\item{csv.file}{Location and name of \code{.csv} file to annotate, either a MoveBank track or file generated by \code{\link{createEnvDataGrid}}}

\item{xml.file}{Location and name of \code{.xml} file that contains the request instructions - almost certainly generated by \code{\link{createEnvDataRequest}}}

\item{link.file}{name of link to access status and final annotated data}

\item{{login, pw}}{login and password for EnvData (NOT the same as MoveBank login and password)}
}
\description{
Upload an Env-Data request for Annotation
}
\details{
The function will fill out the data annotation form on the EnvData annotation tool at \link{http://www.bioinfo.mpg.de/orn-gateway/services2.jsp}.  It will also create a link to the status report of the annotation request, and will attempt to automatically open that page.
}
\examples{
# specify names of data file to annotate (e.g. output of grid) and xml file with the request

csv.file <- "../sandbox/testdata.csv"
xml.file <- "../sandbox/testrequest.xml"

# create a gridded dataset

lats <- seq(38.8, 39.0, length = 40)
lons <- seq(-77.12, -76.91, length = 40) 
start <- ymd("2014-01-01")
finish <- ymd("2014-01-01")

createEnvDataGrid(lats, lons, start, finish, 0, 
                  savefile = TRUE, 
                  fileout = csv.file)

# create an xml request
createEnvDataRequest(request_name = "TestRequest", 
                     type_name = "aster/ASTGTM2", 
                     variable_name = "dem",  
                     interpolation_type = "bilinear",
                     dist_func_spatial="geodetic",
                     email = email,
                     savefile = TRUE, 
                     fileout = xml.file)

# send the request to the server
uploadEnvDataRequest(csv.file, xml.file, 
                     link.file = "annotation_status_link.url")
}
