% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.movetrack.R
\name{map}
\alias{map}
\title{ggmap of a track}
\usage{
map(movetrack, ...)
}
\arguments{
\item{movetrack}{a \code{movetrack} class object, or any data-frame that contains (at least) three columns labeled "time", "x" and "y"}

\item{...}{additional parameters to pass to \code{\link{ggmap}}}

\item{id}{a vector of individuals to map (default = NULL - plots all individuals)}

\item{individually}{whether to produce plots individually (TRUE) or as a population (FALSE; default).}

\item{projection}{one of the options from \link{\code{mapproject}}.  Default - "albers", which seems to be best for  (must be character string).}

\item{print}{whether or not to print map.  Default TRUE}

\item{exportPDF}{a logical indicating whether or not to export plots as a PDF (default = FALSE).}

\item{zoom}{the zoom level - an integer from 3 (continent) to 21 (building) - or (by default) automatically selected as one level lower than the output of \code{\link{calc_zoom}}.}

\item{luminance}{the default colors are richer than the pale ggplot palette which is often hard to see on maps.}
}
\value{
if \code{print = FALSE}, will return the ggmap object, otherwise, returns nothing.  If \code{exportPDF = TRUE} outputs a single PDF of all individuals or a population to the working directory.
}
\description{
This is a wrapper around the \code{\link{ggmap}} function that plots a \code{movetrack} object (including with multiple individuals).
}
\examples{
### Enter Login and Password information for Movebank

## Not run: 
# login <- movebankLogin(username="xxxx", password="xxxx")

# Pull data from movebank
ge1 <- getMovebankData(study="ABoVE: HawkWatch International Golden Eagles", login=login)     #multiple birds

# Estimate daily mean locations
ge1.dailymean <- processMovedata(ge1,  id = "individual_id", dailymean = TRUE)

# Three convenient "movetrack" methods:
summary(ge1.dailymean)
plot(ge1.dailymean)
map(ge1.dailymean)


# one bird, three deployments
ge2 <- getMovebankData(study="Aquila chrysaetos interior west N. America, Craigs, Fuller", animalName="629-26704", login=login) 

# all locations and all columns
ge2.dailymean <- processMovedata(ge2,  id = "individual_id", keepCols = "all")
summary(ge2.dailymean)
plot(ge2.dailymean, id = ge2.dailymean$id[1])
map(ge2.dailymean)
}
