% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMultimigrationFit.R
\name{plotMultiMigration}
\alias{plotMultiMigration}
\title{Plot Multiple Migration Fit}
\usage{
plotMultiMigration(data, M.summary, ...)
}
\arguments{
\item{data}{A processed daily averaged data frame, i.e. output of \code{\link{processMovedata}}}

\item{M.summary}{A data frame with the multi-migration summary (output of \code{\link{fitMultiMigration}})}
}
\description{
This function plots a multiple migration fit using the summary output of \code{\link{fitMultiMigration}}.
}
\examples{
## Not run: 
### Enter Login and Password information for Movebank
# login <- movebankLogin(username="xxxx", password="xxxx")

# Download data
ds <- getMovebankData(study="ABoVE: NPS Dall Sheep Lake Clark", login=login) 

# Process data for use in above package
ds.mt <- processMovedata(ds,  idcolumn = "individual_id", projTo = '+init=epsg:3338', 
                        keepCols = c('sex'), dailymean = TRUE)

# Summarize and find individual
summary(ds.mt)
plot(ds.mt)

#240653478 or 240653490
#di <- subset(ds.mt, id == 240653490) # doesn't maintain class designation
di <- ds.mt[ds.mt$id == 240653490, ]

par(mfrow=c(2,1), mar = c(0,0,0,0), oma = c(5,5,2,2))
with(di, plot(day, x, type='o'))
with(di, plot(day, y, type='o'))

span1 <- locator(4)$x
span2 <- locator(4)$x

di.migrations <- fitMultiMigration(di, span1, span2)
}
