% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickUTM.R
\name{quickUTM}
\alias{quickUTM}
\title{Find UTM zone}
\usage{
quickUTM(lon, lat, singlezone = TRUE, zone = NULL)
}
\arguments{
\item{singlezone}{coerce all data to a single zone;  default TRUE, if FALSE, each location gets its own zone, and the vector of UTM zones is returned.}

\item{zone}{the zone to coerce to - if NULL (default), use the zone of the mid-point of the longitude}

\item{{lon, lat}}{vectors of longitude and latitude.}
}
\value{
data frame with columns X and Y in automatically selected zone.
}
\description{
Uses longitude to identify the UTM zone and converts Long/Lat coords to UTMs.  
This function is buried in \code{\link{processMovedata}}
}
