% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickUTM.R
\name{quickUTM}
\alias{quickUTM}
\title{Find UTM zone}
\usage{
quickUTM(lon, lat)
}
\arguments{
\item{{lon}}{a vector of longitudes.}
}
\value{
data frame with columns X and Y in automatically selected zone.
}
\description{
Uses longitude to identify the UTM zone and converts Long/Lat coords to UTMs.  
This function is buried in \code{\link{processMovedata}}
}
