% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.movetrack.R
\name{summary.movetrack}
\alias{summary.movetrack}
\title{Summary method for a track}
\usage{
\method{summary}{movetrack}(object)
}
\arguments{
\item{object}{a 'movetrack' object, output from \code{\link{processMovedata}}}
}
\value{
a data frame with a column of id's, number of observations, start time, duration and median observation interval of the track.
}
\description{
Summary method for a track
}
\examples{
### Enter Login and Password information for Movebank

## Not run: 
# login <- movebankLogin(username="xxxx", password="xxxx")

# Pull data from movebank
ge1 <- getMovebankData(study="ABoVE: HawkWatch International Golden Eagles", login=login)     #multiple birds

# Estimate daily mean locations
ge1.dailymean <- processMovedata(ge1,  id = "individual_id", dailymean = TRUE)

# Three convenient "movetrack" methods:
summary(ge1.dailymean)
plot(ge1.dailymean)
map(ge1.dailymean)


# one bird, three deployments
ge2 <- getMovebankData(study="Aquila chrysaetos interior west N. America, Craigs, Fuller", animalName="629-26704", login=login) 

# all locations and all columns
ge2.dailymean <- processMovedata(ge2,  id = "individual_id", keepCols = "all")
summary(ge2.dailymean)



# all locations and all columns
ds.df <- processMovedata(ds,  id = "individual_id", keepCols = "all", dailymean=TRUE)
summary(ds.df)
map(ds.df)
}
